FaceHugger Fixes
Tweaked GenericHostileAI to not force MobMeleeAttack, moved code for Face Hugging to its own FaceHugAction script.
