+++ UnityProject/Assets/Scripts/NPC/AI/Hostile/FaceHugAction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a10a3248df701d54a8766c917e03aa80
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
