+++ UnityProject/Assets/Scripts/NPC/AI/Hostile/FaceHuggerAI.cs
@@ -18,21 +18,13 @@ namespace Systems.MobAIs
 		[SerializeField]
 		[Tooltip("If true, this hugger won't be counted for the cap Queens use for lying eggs.")]
 		private bool ignoreInQueenCount = false;
-		[FormerlySerializedAs("Bite")] [SerializeField]
-		private AddressableAudioSource bite = null;
-		[SerializeField] private GameObject maskObject = null;
-		private MobMeleeAction mobMeleeAction;
-
-		protected override void Awake()
-		{
-			base.Awake();
-			simpleAnimal = GetComponent<SimpleAnimal>();
-		}
+		//private MobMeleeAction mobMeleeAction;
+		private FaceHugAction faceHugAction;
 
 		public override void OnEnable()
 		{
 			base.OnEnable();
-			mobMeleeAction = gameObject.GetComponent<MobMeleeAction>();
+			faceHugAction = gameObject.GetComponent<FaceHugAction>();
 		}
 
 		/// <summary>
@@ -73,104 +65,6 @@ namespace Systems.MobAIs
 			}
 		}
 
-		private void TryFacehug(Vector3 dir, LivingHealthMasterBase player)
-		{
-			var playerInventory = player.gameObject.GetComponent<PlayerScript>()?.Equipment;
-
-			if (playerInventory == null)
-			{
-				return;
-			}
-
-			string verb;
-			bool success;
-
-			if (HasAntihuggerItem(playerInventory))
-			{
-				verb = "tried to hug";
-				success = false;
-			}
-			else
-			{
-				verb = "hugged";
-				success = true;
-			}
-
-			mobMeleeAction.ServerDoLerpAnimation(dir);
-
-			Chat.AddAttackMsgToChat(
-				gameObject,
-				player.gameObject,
-				BodyPartType.Head,
-				null,
-				verb);
-
-			AudioSourceParameters audioSourceParameters = new AudioSourceParameters(pitch: 1f);
-			SoundManager.PlayNetworkedAtPos(bite, player.gameObject.RegisterTile().WorldPositionServer,
-				audioSourceParameters, true, player.gameObject);
-
-			if (success)
-			{
-				RegisterPlayer registerPlayer = player.gameObject.GetComponent<RegisterPlayer>();
-				Facehug(playerInventory, registerPlayer);
-			}
-
-		}
-
-		private void Facehug(Equipment playerInventory, RegisterPlayer player)
-		{
-			var result = Spawn.ServerPrefab(maskObject);
-			var mask = result.GameObject;
-
-			Inventory.ServerAdd(
-				mask,
-				playerInventory.ItemStorage.GetNamedItemSlot(NamedSlot.mask),
-				ReplacementStrategy.DespawnOther);
-
-			Despawn.ServerSingle(gameObject);
-		}
-
-		/// <summary>
-		/// Check the player inventory for an item in head, mask or eyes slots with
-		/// Antifacehugger trait. It also drops all items that doesn't have the trait.
-		/// </summary>
-		/// <param name="equipment"></param>
-		/// <returns>True if the player is protected against huggers, false it not</returns>
-		private bool HasAntihuggerItem(Equipment equipment)
-		{
-			bool antiHugger = false;
-
-			foreach (var slot in faceSlots)
-			{
-				var item = equipment.ItemStorage.GetNamedItemSlot(slot)?.Item;
-				if (item == null || item.gameObject == null)
-				{
-					continue;
-				}
-
-				if (!Validations.HasItemTrait(item.gameObject, CommonTraits.Instance.AntiFacehugger))
-				{
-					Inventory.ServerDrop(equipment.ItemStorage.GetNamedItemSlot(slot));
-				}
-				else
-				{
-					var integrity = item.gameObject.GetComponent<Integrity>();
-					if (integrity != null)
-					{
-						// Your protection might break!
-						integrity.ApplyDamage(7.5f, AttackType.Melee, DamageType.Brute);
-					}
-					antiHugger = true;
-				}
-			}
-			return antiHugger;
-		}
-
-		public override void ActOnLiving(Vector3 dir, LivingHealthMasterBase healthBehaviour)
-		{
-			TryFacehug(dir, healthBehaviour);
-		}
-
 		public bool WillInteract(HandApply interaction, NetworkSide side)
 		{
 			return DefaultWillInteract.Default(interaction, side)
@@ -182,7 +76,7 @@ namespace Systems.MobAIs
 		{
 			var handSlot = interaction.HandSlot;
 
-			var result = Spawn.ServerPrefab(maskObject);
+			var result = Spawn.ServerPrefab(faceHugAction.MaskObject);
 			var mask = result.GameObject;
 
 			if (IsDead || IsUnconscious)
@@ -205,12 +99,5 @@ namespace Systems.MobAIs
 			ResetBehaviours();
 			BeginSearch();
 		}
-
-		private readonly List<NamedSlot> faceSlots = new List<NamedSlot>()
-		{
-			NamedSlot.eyes,
-			NamedSlot.head,
-			NamedSlot.mask
-		};
 	}
 }
